move web boilerplate to own module
