+++ compass-geocoder-web-template/src/commonMain/kotlin/dev/jordond/compass/geocoder/web/parameter/TemplateParameters.kt
@@ -1,6 +1,9 @@
 package dev.jordond.compass.geocoder.web.parameter
 
 import dev.drewhamilton.poko.Poko
+import dev.jordond.compass.tools.web.parameter.QueryParameters
+import dev.jordond.compass.tools.web.parameter.QueryParametersBuilder
+import dev.jordond.compass.tools.web.parameter.parametersOf
 
 /**
  * Options for the TEMPLATE geocoder.
