+++ compass-geocoder-web/src/commonMain/kotlin/dev/jordond/compass/geocoder/web/ReverseHttpApiPlatformGeocoder.kt
@@ -3,7 +3,7 @@ package dev.jordond.compass.geocoder.web
 import dev.jordond.compass.Coordinates
 import dev.jordond.compass.Place
 import dev.jordond.compass.exception.NotSupportedException
-import dev.jordond.compass.geocoder.web.internal.makeRequest
+import dev.jordond.compass.tools.web.makeRequest
 import io.ktor.client.HttpClient
 
 /**
