+++ compass-tools-web/src/commonMain/kotlin/dev/jordond/compass/tools/web/exception/WebException.kt
@@ -1,0 +1,8 @@
+package dev.jordond.compass.tools.web.exception
+
+import io.ktor.http.HttpStatusCode
+
+public class WebException(
+    public val statusCode: HttpStatusCode,
+    public override val message: String,
+) : Throwable(message)
