+++ compass-tools-web/src/commonMain/kotlin/dev/jordond/compass/tools/web/parameter/QueryParamValue.kt
@@ -1,4 +1,4 @@
-package dev.jordond.compass.geocoder.web.parameter
+package dev.jordond.compass.tools.web.parameter
 
 /**
  * Represents a query parameter value.
