+++ compass-tools-web/src/commonMain/kotlin/dev/jordond/compass/tools/web/parameter/QueryParameters.kt
@@ -1,4 +1,4 @@
-package dev.jordond.compass.geocoder.web.parameter
+package dev.jordond.compass.tools.web.parameter
 
 import io.ktor.http.encodeURLQueryComponent
